[Dropdown]: improves styling for dropdown popup width
Previously, the popup would expand to the width of the control
and the label, if the label was horizontally aligned. This ensures
that the popup width is based on the control, and not the label.

Additionally exposes the control width to the consumer via the
--leo-menu-control-width variable, and optionally allows the consumer to
set the max-width via --leo-menu-popup-max-width.