[Tokens Parser]: changes how token name segments are removed
Prior to this commit, we removed token segments via string manipulation
with basic pattern find/replace manipulation. The problem with this
method is that we don't have a means of distinguishing between, say,
`--leo-font-browser-large-regular` (where we want to remove `-browser-`) and
`--leo-color-chrome-browser-stroke` (where we do not want to remove
`-browser-`). This commit fixes the problem by removing the key from the
object during parsing.