[Tokens]: update existing tokens and add more layers  (#543)
Token parser updates:
- add more shadows to be grouped like 'focus state'
- add support for layer specific effect colors
- change how token name segments are removed
Prior to this commit, we removed token segments via string manipulation
with basic pattern find/replace manipulation. The problem with this
method is that we don't have a means of distinguishing between, say,
`--leo-font-browser-large-regular` (where we want to remove `-browser-`) and
`--leo-color-chrome-browser-stroke` (where we do not want to remove
`-browser-`). This commit fixes the problem by removing the key from the
object during parsing.

Tokens updates:
- add 'android' and 'ios' to be built for web platforms
- add 'browser' layer to android, ios, and skia builds
- update universal tokens/variables
- update browser tokens
- update android tokens
- add marketing tokens
- add search tokens
- add news tokens
- add newtab tokens
- add web3 tokens

---------

Co-authored-by: Jacob Lamont <cob@jacoblamont.com>